backend-migration-job/requirements.txt
PyGithub==2.5.0
GitPython==3.1.43
requests==2.32.3
